lua: rename {utils,websocket}.lua->firenvim{utils,websocket}.lua
